requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
googletrans==4.0.0rc1
openai==0.27.9
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
PyMuPDF==1.23.3 